feat(macos,#759): XR_DXR_weave service on the macOS shared-surface path
Port the window-bound synchronous weave service (#625, spec v3 batch #744)
to the macOS OOP service: present-owner callers (the future mac DisplayXR
Browser GPU process) hand the runtime pre-weave SBS IOSurface pixels +
window-relative rects and composite back a weaved IOSurface.

- comp_multi_weave_macos.c: per-client Vulkan weave engine on the null
  compositor's vk_bundle — IOSurface input import (VK_EXT_metal_objects +
  VK_EXT_external_memory_metal), window-sized 2x1 SBS scratch, per-rect
  vkCmdBlitImage, ONE process_atlas per submit (dp_factory_vk, the same DP
  family the shared surface drives; sim_display anaglyph runs on MoltenVK),
  IOSurface-backed output export. Completion is SYNCHRONOUS (vkWaitForFences
  before the IPC reply; no cross-process fence on macOS).
- ipc_server_handler.c: XRT_OS_MACOS branches in the weave handlers
  (bind/submit/get_output/snap; get_fence keeps the no-fence fallback).
- oxr_extension_support.h: widen the OXR_HAVE_DXR_weave gate to
  XR_USE_PLATFORM_MACOS; oxr_weave.c handles fd-typed sync handles.
- Output sizing: batch = input IOSurface dims (v3 input is window-client-
  sized by contract); legacy = rect offset+extent. Bind stores the id only;
  snap is identity (no VK DP snap slot yet).
- test_apps/probes/weave_probe_vk_macos: headless present-owner harness
  (forced-IPC MoltenVK session, CPU-filled SBS IOSurface, batched submit,
  CPU-verifies the anaglyph weave). PASSES against sim_display anaglyph:
  left-eye-white rect weaves to (255,0,0), right-eye-white to (0,255,255).
- docs/specs/extensions/XR_DXR_weave.md: macOS platform-mapping section.

+++ src/xrt/state_trackers/oxr/oxr_weave.c
@@ -14,7 +14,9 @@
  * (ADR-007 / ADR-019).
  *
  * Availability: implemented only on the out-of-process (service / IPC) path —
- * the weave runs in the D3D11 service compositor. An in-process session reports
+ * the weave runs in the D3D11 service compositor on Windows and in the
+ * comp_multi Vulkan weave engine on macOS (#759: IOSurface in/out, synchronous
+ * completion, no fence handle). An in-process session reports
  * XR_ERROR_FEATURE_UNSUPPORTED. The entry points forward to thin IPC-client
  * bridges (defined in ipc_client_compositor.c); st_oxr does not pull the
  * ipc_client include path, so the symbols resolve at link time — same pattern
@@ -227,7 +229,10 @@ oxr_xrWeaveSubmitDXR(XrSession session, const XrWeaveSubmitInfoDXR *submitInfo, 
 		if (comp_ipc_client_compositor_weave_get_fence(&sess->xcn->base, &have_fence, &fence_h) ==
 		        XRT_SUCCESS &&
 		    have_fence && fence_h != XRT_GRAPHICS_SYNC_HANDLE_INVALID) {
-			output->fence = (void *)fence_h;
+			// (intptr_t hop: the handle is an fd int on POSIX — macOS never
+			// exports a fence (#759, completion is synchronous), so this
+			// branch fires on Windows HANDLEs only.)
+			output->fence = (void *)(intptr_t)fence_h;
 		}
 
 		sess->weave.exported = true;
